chess-bot/requirements.txt
numpy
torch
flask
transformers
faiss-cpu
PyPDF2
flask-cors
boto3
python-dotenv
gunicorn